Move the ProcessCleanAction method into the base, so any handler can do it.
